update to version 1.2.4
+ core changes
+ fixed bugs
+ new help
+ add softban

+++ commands/kick.js
@@ -1,57 +1,59 @@
 const Discord = module.require("discord.js");
 
 module.exports.execute = async (bot, message, content, config, moment, request) => {
+    let errorEmbed = new Discord.RichEmbed()
+        .setColor(`#${config.colorDanger}`)
+        .setDescription(`**<@${message.author.id}>, ${config.noPermission}**`);
+    let usageEmbed = new Discord.RichEmbed()
+        .setColor(`#${config.colorInfo}`)
+        .setTitle(`Correct usage for ${config.prefix}${this.help.name} command.`)
+        .setDescription(`\`${config.prefix}${this.help.name} ${this.help.usage}\` - *${this.help.description}*`);
+    let notKickableEmbed = new Discord.RichEmbed()
+        .setColor(`#${config.colorDanger}`)
+        .setDescription(`**<@${message.author.id}>, I can't kick this user! Do they have a higher role?**`);
+
     if (!message.member.hasPermission("KICK_MEMBERS")) {
-        let errorEmbed = new Discord.RichEmbed()
-            .setColor(`#${config.colorDanger}`)
-            .setDescription(`**<@${message.author.id}>, ${config.noPermission}**`)
-        await message.channel.send(errorEmbed);
+        await message.channel.send(errorEmbed).catch(O_o => {});
         return;
     }
 
-    let helpName = this.help.name;
-    let helpUsage = this.help.usage;
-    let helpDescription = this.help.description;
-    try {
-        let member = message.mentions.members.first() || message.guild.members.get(content[0]);
-        if (!member.kickable) {
-            let errorEmbed = new Discord.RichEmbed()
-                .setColor(`#${config.colorDanger}`)
-                .setDescription(`**<@${message.author.id}>, I can't kick this user! Do they have a higher role?**`)
-            await message.channel.send(errorEmbed);
-            return;
-        }
+    if (content[0]) {
+        let id = content[0].replace(/[^a-zA-Z0-9]/g, '');
+        try {
+            let member = message.guild.members.get(id);
+            if (!member.kickable) {
+                await message.channel.send(notKickableEmbed).catch(O_o => {});
+                return;
+            }
 
-        let reason = content.slice(1).join(' ');
-        if (!reason) reason = "Not following our rules!";
-        await message.delete().catch(O_o => {});
-        await member.kick(reason).catch(O_o => {});
-        let logEmbed = new Discord.RichEmbed()
+            let reason = content.slice(1).join(' ');
+            if (!reason) reason = "None";
+            await message.delete().catch(O_o => {});
+            await member.kick(reason).catch(O_o => {});
+            let logEmbed = new Discord.RichEmbed()
                 .setColor(`#${config.colorInfo}`)
                 .setAuthor(`Kicked | ${member.user.tag}`, bot.user.displayAvatarURL)
-                .setThumbnail(bot.user.displayAvatarURL)
+                .setThumbnail(config.securityShield)
                 .addField("User", member.user.tag, true)
                 .addField("Staff Member", message.author.tag, true)
                 .addField("Reason", reason)
                 .setFooter(`${moment.tz(message.createdTimestamp, config.timezone).format(config.timeFormat)}`);
-        if (message.guild.id === config.officialGuildID) {
-            await bot.channels.get(config.logChannelID).send(logEmbed).catch(O_o => {});
-        } else {
+            if (message.guild.id === config.officialGuildID) {
+                await bot.channels.get(config.logChannelID).send(logEmbed).catch(O_o => {});
+                return;
+            } 
             await message.channel.send(logEmbed).catch(O_o => {});
-        }
-    } catch(e) {
-        let usageEmbed = new Discord.RichEmbed()
-            .setColor(`#${config.colorInfo}`)
-            .setTitle(`Correct usage for ${config.prefix}${helpName} command.`)
-            .setDescription(`\`${config.prefix}${helpName} ${helpUsage}\` - *${helpDescription}*`);
+        } catch(e) { await message.channel.send(usageEmbed).catch(O_o => {}); }
+        return;
+    } else {
         await message.channel.send(usageEmbed).catch(O_o => {});
+        return;
     }
-    return;
 }
 
 module.exports.help = {
     name: "kick",
-    usage: "[user/user id] <reason>",
+    usage: ["[user] <reason>"],
     category: "admin",
     description: "Kick a user."
 }
